pyproject.toml
[tool.poetry]
name = "showerthoughtsdino"
version = "0.1.0"
description = "A bot that gets posts from r/showerthougths and posts it to Slack."
authors = ["Arpan Pandey <arpan@hackersreboot.tech>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
slack-bolt = "^1.16.1"
python-dotenv = "^0.21.1"
praw = "^7.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
